requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
decorator==4.4.2
distlib==0.3.0
docopt==0.6.2
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
Js2Py==0.70
MouseInfo==0.1.3
moviepy==1.0.3
numpy==1.18.5
opencv-python==4.2.0.34
packaging==20.4
Pillow==7.1.2
pipwin==0.5.0
proglog==0.1.9
PyAudio @ file:///C:/Users/dst11/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
PyAutoGUI==0.9.50
pydub==0.24.1
PyGetWindow==0.0.8
pyjsparser==2.7.1
PyMsgBox==1.0.8
pynput==1.6.8
pyparsing==2.4.7
pyperclip==1.8.0
PyPrind==2.11.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.13.2.1.6rc1
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.3
python-dotenv==0.13.0
PyTweening==1.0.3
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.23
